+++ satella/coding/concurrent/__init__.py
@@ -1,6 +1,6 @@
-from .callablegroup import *
+from .callablegroup import CallableGroup
 from .locked_dataset import LockedDataset
-from .monitor import *
+from .monitor import MonitorList, Monitor, MonitorDict, RMonitor
 from .thread import TerminableThread, Condition
 
 __all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',