+++ satella/coding/structures/__init__.py
@@ -3,9 +3,11 @@ from .immutable import Immutable, frozendict
 from .singleton import Singleton, SingletonWithRegardsTo
 from .structures import SetHeap, OmniHashableMixin, TimeBasedHeap, Heap
 from .typednamedtuple import typednamedtuple
+from .hashable_objects import HashableWrapper
 
 __all__ = [
     'typednamedtuple',
+    'HashableWrapper',
     'Heap',
     'SetHeap',
     'DictionaryView',